+++ db/migrate/20181226173221_create_transactions.rb
@@ -2,7 +2,7 @@ class CreateTransactions < ActiveRecord::Migration[5.2]
   def change
     create_table :transactions do |t|
       t.boolean :retrait
-      t.integer :montant
+      t.float :montant
       t.boolean :automatique
       t.references :cagnotte, foreign_key: true
